Update package requirements, especially Biopython to 1.78
Code updated to cope with the removal of the Bio.Alphabet module in
Biopython 1.78 (September 2020). The remaining requirements have
been updated too to current versions after comprehensive testing.

NOTE: The functionality remains the same as in the previous version
(1.1.2) so, if you need to work with a Biopython release preceding
v1.78 you can keep working with Recentrifuge 1.1.2 by now.

On branch master
	modified:   rcf
	modified:   recentrifuge/__init__.py
	modified:   recentrifuge/centrifuge_io.py
	modified:   recentrifuge/core.py
	modified:   recentrifuge/fastq_io.py
	modified:   recentrifuge/lmat_io.py
	modified:   recentrifuge/mock.py
	modified:   recentrifuge/rank.py
	modified:   recentrifuge/taxclass.py
	modified:   recentrifuge/taxonomy.py
	modified:   recentrifuge/trees.py
	modified:   requirements.txt
	modified:   setup.py